Refactor javascript file
